web自动化/requirement.txt
pytest
allure-pytest
pytest-html
pytest-xdist